% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData.R
\name{simData}
\alias{simData}
\title{Generate simulated data for \code{\link{CHiMiC}} and \code{\link{CV.CHiMiC}}}
\usage{
simData(n, s1 = 20, s2 = 5, ratio = 0.2)
}
\arguments{
\item{n}{An integer specifying the number of samples to generate.}

\item{s1}{An integer specifying the number of relevant continuous variables.}

\item{s2}{An integer specifying the number of relevant categorical variables.}

\item{ratio}{A numeric value between 0 and 1 specifying the proportion of relevant variables among the total continuous/categorical variables considered.}
}
\value{
A "simData" object containing the following components:
\item{Z}{ A simulated matrix of clinical covariates with rows representing samples and columns representing variables (q1 + q2 in total).}
\item{q1}{ An integer indicating that the first q1 columns of Z are continuous covariates.}
\item{q2}{ An integer indicating that the last q2 columns of Z are categorical covariates.}
\item{CATE}{ A simulated numeric vector used to guide clustering.}
\item{G}{ Number of clusters.}
\item{clusters}{ A numeric vector of cluster assignments.}
\item{features_ind}{ Indices of relevant variables in Z.}
}
\description{
Generate simulated data for applying \code{\link{CHiMiC}} and \code{\link{CV.CHiMiC}}.
}
\examples{
set.seed(123)
simdata <- simData(n = 1000, s1 = 20, s2 = 5, ratio = 0.2)
Z <- simdata$Z
G <- simdata$G
q1 <- simdata$q1
q2 <- simdata$q2
cate <- simdata$CATE
hist(cate, breaks = 200)
}
\seealso{
See Also as \code{\link{CHiMiC}}, \code{\link{CV.CHiMiC}}.
}
